moved tests to describeConformance
