+++ test/utils/describeConformance.js
@@ -47,7 +47,7 @@ export function findRootComponent(wrapper, { component }) {
   });
 }
 
-export function randomStringValue() {
+function randomStringValue() {
   return `s${Math.random().toString(36).slice(2)}`;
 }
 
@@ -148,7 +148,7 @@ export function testRootClass(element, getOptions) {
     }
 
     const className = randomStringValue();
-    const wrapper = mount(React.cloneElement(element, { className }));
+    let wrapper = mount(React.cloneElement(element, { className }));
 
     // we established that the root component renders the outermost host previously. We immediately
     // jump to the host component because some components pass the `root` class
@@ -156,9 +156,16 @@ export function testRootClass(element, getOptions) {
     // https://github.com/mui-org/material-ui/blob/f9896bcd129a1209153106296b3d2487547ba205/packages/material-ui/src/OutlinedInput/OutlinedInput.js#L101
     expect(findOutermostIntrinsic(wrapper).hasClass(classes.root)).to.equal(true);
     expect(findOutermostIntrinsic(wrapper).hasClass(className)).to.equal(true);
+
+    // Test that classes prop works
+    const classesProp = { ...classes };
+    classesProp.root = `${classesProp.root} ${className}`;
+
+    wrapper = mount(React.cloneElement(element, { classes: classesProp }));
+    expect(findOutermostIntrinsic(wrapper).hasClass(className)).to.equal(true);
+    expect(findOutermostIntrinsic(wrapper).getDOMNode().getAttribute('classes')).to.equal(null);
   });
 }
-
 /**
  * Tests that the component can be rendered with react-test-renderer.
  * This is important for snapshot testing with Jest (even if we don't encourage it).
